requirements.txt
python-telegram-bot>=20.0
langchain-community>=0.0.10
ollama>=0.1.0